Refactor action dialog data source and add force count action
The action dialog used to pull the source/target data itself.
But for some actions that was not working well.

Now the action dialog is given the information during the setup.
This fixes label / unlabel and allows better local actions.

New action Force count is added that allows updating the location
count according to the real world count done by the user.