Reimplementation of the select item logic
There are now three different ways to select a panel item.
By name, by part id and by display id.

Display id is a hash of the part id and its parent id and
is used after reload to select the same item that was selected
before the reload. In the past that was done by name only,
however that is insufficient when multiple sized pieces
of the same part show up.

For the same reason the select by PartId was added when creating
new items.

Select by name was kept for search purposes.